backend/requirements/base.txt
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
sqlalchemy>=2.0.43,<3.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
strawberry-graphql
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
email-validator>=2.2.0
python-dotenv==1.0.0
